Extract CT-RAMP pre-formatting filters into filters.py
format_ctramp.py mixed the @step orchestrator with ~290 lines of record
filtering that clutter it. Move the two drop concerns to a sibling
filters.py, matching the module's one-concern-per-file layout
(format_households/persons/tours/trips):

- _drop_invalid_tours (+ _drop_by_tour_ids, _completeness_split,
  _drop_grid, _log_drop_summary): invalid/partial tour drop + drop log
- _drop_missing_taz: TAZ filtering cascade

Pure move, names unchanged; format_ctramp.py imports the two entry
points. Also:
- Dedupe the repeated "is_not_null() & != -1" TAZ predicate in
  _drop_missing_taz behind a _valid_taz(field) helper (equivalence).
- Fix _drop_excess_fields docstring: drop a stale copy of an old
  implementation that had been pasted inside the triple-quotes.
- Point the one external import (test_ctramp_integration) at filters.

format_ctramp.py 799 -> 505 lines. 850 tests pass, lint + format clean.
