+++ src/processing/formatting/ctramp/format_ctramp.py
@@ -194,7 +194,6 @@ import logging
 import polars as pl
 
 from data_canon.codebook.ctramp import CTRAMPEmploymentCategory
-from data_canon.codebook.tours import TourCategory, TourDataQuality
 from data_canon.models.ctramp import (
     AOResultsCTRAMPModel,
     CDAPResultsCTRAMPModel,
@@ -209,6 +208,7 @@ from data_canon.models.ctramp import (
 from pipeline.decoration import step
 
 from .ctramp_config import CTRAMPConfig
+from .filters import _drop_invalid_tours, _drop_missing_taz
 from .format_ao import format_ao_results
 from .format_cdap import format_cdap_results
 from .format_households import format_households
@@ -234,298 +234,6 @@ MODEL_MAP = {
 }
 
 
-def _drop_by_tour_ids(
-    tours: pl.DataFrame,
-    linked_trips: pl.DataFrame,
-    joint_trips: pl.DataFrame,
-) -> tuple[pl.DataFrame, pl.DataFrame, pl.DataFrame]:
-    """Cascade a tour filter down to linked trips and joint trips.
-
-    Keeps only linked trips whose ``tour_id`` still exists in ``tours``, then
-    keeps only joint trips whose ``joint_trip_id`` still appears in the surviving
-    linked trips. This preserves referential integrity after tours are removed.
-
-    Args:
-        tours: Tours that survived a filter (defines the surviving tour_ids)
-        linked_trips: Linked trips to prune to the surviving tours
-        joint_trips: Joint trips to prune to the surviving linked trips
-
-    Returns:
-        Tuple of (tours, linked_trips, joint_trips) with integrity maintained
-    """
-    linked_trips = linked_trips.filter(pl.col("tour_id").is_in(tours["tour_id"].implode()))
-
-    if len(joint_trips) > 0 and "joint_trip_id" in linked_trips.columns:
-        valid_joint_trip_ids = linked_trips.filter(pl.col("joint_trip_id").is_not_null())[
-            "joint_trip_id"
-        ].unique()
-        joint_trips = joint_trips.filter(
-            pl.col("joint_trip_id").is_in(valid_joint_trip_ids.implode())
-        )
-
-    return tours, linked_trips, joint_trips
-
-
-def _drop_invalid_tours(
-    tours: pl.DataFrame,
-    linked_trips: pl.DataFrame,
-    joint_trips: pl.DataFrame,
-) -> tuple[pl.DataFrame, pl.DataFrame, pl.DataFrame]:
-    """Remove tours unusable for the model, cascading to linked and joint trips.
-
-    Drops tours that are not VALID (single-trip, loop, missing-anchor,
-    change-mode, indeterminate) or not COMPLETE (do not start and end at home).
-    Mirrors the DaySim formatter (``format_daysim``), which drops both, so the
-    CT-RAMP and DaySim outputs contain the same set of tours. Without this,
-    single-trip/loop tours (which carry a null ``tour_purpose``) survive into
-    CT-RAMP output and are silently coerced to the ``OTHDISCR`` catch-all purpose.
-
-    Args:
-        tours: Canonical tour data with ``tour_data_quality`` and ``tour_category``
-        linked_trips: Canonical linked trip data
-        joint_trips: Aggregated joint trip data
-
-    Returns:
-        Tuple of (tours, linked_trips, joint_trips) with unusable tours and their
-        orphaned trips removed.
-    """
-    # tour_data_quality is produced during tour extraction; guard for callers
-    # (e.g. schema-only empty frames) that do not carry every column.
-    has_quality = "tour_data_quality" in tours.columns
-    has_category = "tour_category" in tours.columns
-    if not (has_quality or has_category):
-        return tours, linked_trips, joint_trips
-
-    is_valid = pl.col("tour_data_quality") == TourDataQuality.VALID.value
-    is_complete = pl.col("tour_category") == TourCategory.COMPLETE.value
-
-    # Tag each tour's drop reason so the log can distinguish structurally invalid
-    # tours from partial/open ones (valid but not home-based).
-    if has_quality and has_category:
-        keep = is_valid & is_complete
-        reason = pl.when(~is_valid).then(pl.lit("invalid")).otherwise(pl.lit("partial/open"))
-    elif has_quality:
-        keep = is_valid
-        reason = pl.lit("invalid")
-    else:
-        keep = is_complete
-        reason = pl.lit("partial/open")
-
-    n_og_tours = len(tours)
-    dropped = tours.filter(~keep).with_columns(reason.alias("_reason"))
-    _log_drop_summary(dropped, linked_trips, n_og_tours)
-
-    tours = tours.filter(keep)
-    tours, linked_trips, joint_trips = _drop_by_tour_ids(tours, linked_trips, joint_trips)
-    return tours, linked_trips, joint_trips
-
-
-def _completeness_split(df: pl.DataFrame) -> tuple[int, int]:
-    """Return (already_incomplete, net_new) counts from the ``complete`` flag.
-
-    ``already_incomplete`` are records flagged incomplete (household / person /
-    day cascade) — they were already excluded from the weighted model. ``net_new``
-    are otherwise-complete records now being dropped. When the ``complete`` column
-    is absent (e.g. tests), everything is treated as net-new.
-    """
-    total = df.height
-    if "complete" not in df.columns:
-        return 0, total
-    incomplete = df.filter(~pl.col("complete").fill_null(value=False)).height
-    return incomplete, total - incomplete
-
-
-def _drop_grid(title: str, df: pl.DataFrame, denom: int) -> list[str]:
-    """Build the lines of one drop grid (tours or member trips).
-
-    ``net %`` is the net-new share of all records — the genuinely new data loss.
-    """
-    header = f"  {title:<14}{'dropped':>9}{'incomplete':>13}{'net-new':>10}{'net %':>8}"
-
-    def row(name: str, incomplete: int, net: int) -> str:
-        total = incomplete + net
-        pct = net * 100 / denom if denom > 0 else 0
-        return f"  {name:<14}{total:>9,}{incomplete:>13,}{net:>10,}{pct:>7.1f}%"
-
-    lines = [header]
-    for label in ("invalid", "partial/open"):
-        incomplete, net = _completeness_split(df.filter(pl.col("_reason") == label))
-        if incomplete + net == 0:
-            continue
-        lines.append(row(label, incomplete, net))
-    incomplete, net = _completeness_split(df)
-    lines.append(row("total", incomplete, net))
-    return lines
-
-
-def _log_drop_summary(dropped: pl.DataFrame, linked_trips: pl.DataFrame, n_og_tours: int) -> None:
-    """Log a two-grid table of dropped tours and member trips.
-
-    Splits each count into records that were already incomplete (household /
-    person / day flagged incomplete — already out of the weighted model) versus
-    net-new removal of otherwise-complete records, so the log makes clear how
-    much is genuinely new loss.
-    """
-    if dropped.height == 0:
-        return
-
-    dropped_trips = linked_trips.join(
-        dropped.select("tour_id", "_reason"), on="tour_id", how="inner"
-    )
-    lines = [
-        "Dropped tours/member-trips to match DaySim (DaySim already drops these).",
-        '  "incomplete" = flagged incomplete via the household>person>day>trip cascade; '
-        '"net-new" = otherwise-complete records.',
-        "",
-        *_drop_grid("TOURS", dropped, n_og_tours),
-        "",
-        *_drop_grid("MEMBER TRIPS", dropped_trips, len(linked_trips)),
-    ]
-    logger.info("\n".join(lines))
-
-
-def _drop_missing_taz(
-    households: pl.DataFrame,
-    persons: pl.DataFrame,
-    tours: pl.DataFrame,
-    linked_trips: pl.DataFrame,
-    joint_trips: pl.DataFrame,
-    config: CTRAMPConfig,
-) -> tuple[pl.DataFrame, pl.DataFrame, pl.DataFrame, pl.DataFrame, pl.DataFrame]:
-    """Remove records with missing TAZ fields and ensure referential integrity.
-
-    Performs cascading deletions to maintain data consistency:
-    1. Remove households without valid home TAZ
-    2. Remove persons from dropped households
-    3. Remove tours/trips with missing origin or destination TAZ
-    4. Remove tours that have no trips remaining
-    5. Remove joint trips that have no linked trips remaining
-
-    Args:
-        households: Canonical household data
-        persons: Canonical person data
-        tours: Canonical tour data
-        linked_trips: Canonical linked trip data
-        joint_trips: Canonical joint trip data
-        config: CTRAMP configuration with taz_field name
-
-    Returns:
-        Tuple of filtered DataFrames maintaining referential integrity
-    """
-    # Track original counts for logging
-    counts = {
-        "households": len(households),
-        "persons": len(persons),
-        "tours": max(0, len(tours)),
-        "linked_trips": max(0, len(linked_trips)),
-        "joint_trips": max(0, len(joint_trips)),
-    }
-
-    # Step 1: Filter households by home TAZ
-    households = households.filter(
-        pl.col(f"home_{config.taz_field}").is_not_null()
-        & (pl.col(f"home_{config.taz_field}") != -1)
-    )
-    valid_hh_ids = households["hh_id"]
-
-    # Step 2: Remove orphaned persons
-    persons = persons.filter(pl.col("hh_id").is_in(valid_hh_ids.implode()))
-
-    logger.info(
-        "Dropped %d households without valid home TAZ (keeping %d households, %d persons)",
-        counts["households"] - len(households),
-        len(households),
-        len(persons),
-    )
-
-    # Step 3: Filter tours by household and TAZ fields
-    if len(tours) > 0:
-        tours = tours.filter(
-            pl.col("hh_id").is_in(valid_hh_ids.implode())
-            & pl.col(f"o_{config.taz_field}").is_not_null()
-            & (pl.col(f"o_{config.taz_field}") != -1)
-            & pl.col(f"d_{config.taz_field}").is_not_null()
-            & (pl.col(f"d_{config.taz_field}") != -1)
-        )
-        valid_tour_ids = tours["tour_id"]
-    else:
-        valid_tour_ids = pl.Series("tour_id", [], dtype=pl.Int64)
-
-    # Step 4: Filter linked trips by tour and TAZ fields
-    if len(linked_trips) > 0:
-        linked_trips = linked_trips.filter(
-            pl.col("tour_id").is_in(valid_tour_ids.implode())
-            & pl.col(f"o_{config.taz_field}").is_not_null()
-            & (pl.col(f"o_{config.taz_field}") != -1)
-            & pl.col(f"d_{config.taz_field}").is_not_null()
-            & (pl.col(f"d_{config.taz_field}") != -1)
-        )
-        # Get tours that still have trips
-        tours_with_trips = linked_trips["tour_id"].unique()
-
-        # Remove tours that lost all their trips
-        if len(tours) > 0:
-            tours_before = len(tours)
-            tours = tours.filter(pl.col("tour_id").is_in(tours_with_trips.implode()))
-            if tours_before != len(tours):
-                logger.info(
-                    "Removed %d tours that had no valid trips remaining",
-                    tours_before - len(tours),
-                )
-    # No trips means no tours should remain
-    elif len(tours) > 0:
-        logger.info("Removed all %d tours because no valid trips exist", len(tours))
-        tours = tours.head(0)
-
-    # Step 5: Filter joint trips by linked trips and TAZ fields
-    if len(joint_trips) > 0:
-        # First filter by household
-        joint_trips = joint_trips.filter(pl.col("hh_id").is_in(valid_hh_ids))
-
-        # Filter by TAZ fields if they exist in joint_trips
-        taz_cols = [f"o_{config.taz_field}", f"d_{config.taz_field}"]
-        has_taz = all(col in joint_trips.columns for col in taz_cols)
-        if has_taz:
-            joint_trips = joint_trips.filter(
-                pl.col(f"o_{config.taz_field}").is_not_null()
-                & (pl.col(f"o_{config.taz_field}") != -1)
-                & pl.col(f"d_{config.taz_field}").is_not_null()
-                & (pl.col(f"d_{config.taz_field}") != -1)
-            )
-
-        # Keep only joint trips that have corresponding linked trips
-        if len(linked_trips) > 0 and "joint_trip_id" in linked_trips.columns:
-            valid_joint_trip_ids = linked_trips.filter(pl.col("joint_trip_id").is_not_null())[
-                "joint_trip_id"
-            ].unique()
-            joint_trips = joint_trips.filter(pl.col("joint_trip_id").is_in(valid_joint_trip_ids))
-        else:
-            # No linked trips with joint_trip_id means no joint trips should remain
-            logger.info(
-                "Removed all %d joint trips because no valid linked trips exist", len(joint_trips)
-            )
-            joint_trips = joint_trips.head(0)
-
-    # Final logging
-    logger.info(
-        "TAZ filtering complete:\n"
-        "  Tours: %d → %d (dropped %d)\n"
-        "  Linked trips: %d → %d (dropped %d)\n"
-        "  Joint trips: %d → %d (dropped %d)",
-        counts["tours"],
-        len(tours),
-        counts["tours"] - len(tours),
-        counts["linked_trips"],
-        len(linked_trips),
-        counts["linked_trips"] - len(linked_trips),
-        counts["joint_trips"],
-        len(joint_trips),
-        counts["joint_trips"] - len(joint_trips),
-    )
-
-    return households, persons, tours, linked_trips, joint_trips
-
-
 def _drop_excess_fields(
     df: pl.DataFrame,
     model_cls: type,
@@ -535,11 +243,9 @@ def _drop_excess_fields(
     Args:
         df: Input DataFrame
         model_cls: Data model class with defined fields
+
     Returns:
         DataFrame with only columns defined in the model class
-    valid_fields = set(model_cls.model_fields.keys())
-    cols_to_drop = [col for col in df.columns if col not in valid_fields]
-    return df.drop(cols_to_drop)
     """
     valid_fields = set(model_cls.model_fields.keys())
     cols_to_drop = set(df.columns) - valid_fields
